backend/pipelines/climate_etl.py
"""INMET climate ETL for Nova Corrente - Salvador/BA climate data"""
from typing import Dict, Any, Optional
from datetime import date, datetime, timedelta
import pandas as pd
import requests
import time

from backend.services.database_service import db_service
from backend.config.external_apis_config import INMET_CONFIG
from backend.config.logging_config import get_logger

logger = get_logger('nova_corrente.etl.climate')


class ClimateETL:
    """ETL for INMET climate data"""
    
    def extract(self, start_date: date, end_date: date) -> pd.DataFrame:
        """Extract climate data from INMET API"""
        try:
            # Placeholder for INMET API integration
            # In production, would call INMET API
            logger.info(f"Extracting climate data from {start_date} to {end_date}")
            return pd.DataFrame()
        except Exception as e:
            logger.error(f"Error extracting climate data: {e}")
            raise
    
    def transform(self, raw_data: pd.DataFrame) -> pd.DataFrame:
        """Transform climate data"""
        try:
            if raw_data.empty:
                return pd.DataFrame()
            
            # Calculate derived features
            if 'temperatura_media' in raw_data.columns:
                raw_data['is_extreme_heat'] = raw_data['temperatura_media'] > 35
                raw_data['is_cold_weather'] = raw_data['temperatura_media'] < 20
            
            if 'precipitacao_mm' in raw_data.columns:
                raw_data['is_heavy_rain'] = raw_data['precipitacao_mm'] > 50
                raw_data['is_intense_rain'] = raw_data['precipitacao_mm'] > 100
                raw_data['is_no_rain'] = raw_data['precipitacao_mm'] == 0
            
            if 'umidade_percentual' in raw_data.columns:
                raw_data['is_high_humidity'] = raw_data['umidade_percentual'] > 80
            
            # Calculate risks
            raw_data['corrosion_risk'] = (
                (raw_data.get('umidade_percentual', 0) / 100) * 
                (raw_data.get('precipitacao_mm', 0) / 100)
            ).fillna(0)
            
            raw_data['field_work_disruption'] = (
                (raw_data.get('precipitacao_mm', 0) > 50).astype(int) * 0.5 +
                (raw_data.get('temperatura_media', 0) > 35).astype(int) * 0.3
            ).fillna(0)
            
            return raw_data
        except Exception as e:
            logger.error(f"Error transforming climate data: {e}")
            raise
    
    def load(self, transformed_data: pd.DataFrame) -> int:
        """Load climate data into database"""
        try:
            if transformed_data.empty:
                return 0
            
            rows_inserted = db_service.insert_dataframe(
                'ClimaSalvador',
                transformed_data,
                if_exists='replace'
            )
            
            logger.info(f"Loaded {rows_inserted} climate records")
            return rows_inserted
        except Exception as e:
            logger.error(f"Error loading climate data: {e}")
            raise
    
    def run(self, start_date: date, end_date: date) -> int:
        """Run complete ETL pipeline"""
        try:
            raw_data = self.extract(start_date, end_date)
            transformed_data = self.transform(raw_data)
            rows_inserted = self.load(transformed_data)
            return rows_inserted
        except Exception as e:
            logger.error(f"Error running climate ETL: {e}")
            raise


climate_etl = ClimateETL()

